com.unity.render-pipelines.core@5.16.1 ## [5.16.1] - 2019-05-22 ### Changed - This package now requires Unity 2019.1.3f1 or later to run.
## [5.16.0] - 2019-05-20

### Version Updated
The version number for this package has increased due to a version update of a related graphics package.

## [5.15.0] - 2019-05-13
### Changed
The version number for this package has increased due to a version update of a related graphics package.

## [5.14.0] - 2019-05-09

### Changed
The version number for this package has increased due to a version update of a related graphics package.